requirements.txt
Flask==2.2.5
Flask_PyMongo==2.3.0
googlemaps==4.10.0
pydantic==2.5.2
pymongo==4.5.0
python-dotenv==0.21.1
Requests==2.31.0
Gunicorn==21.2.0
